refactor: Layoutコンポーネントを/page下に移動
