+++ src/components/page/layout/Layout.stories.tsx
@@ -5,7 +5,7 @@ import { Layout } from "./Layout";
 import type { Meta } from "@storybook/react";
 
 const meta: Meta<typeof Layout> = {
-  title: "Layout",
+  title: "Pages/Layout",
   component: Layout,
 };
